Enable fail-fast on the gate queue
Zuul recently added fail-fast[1] as a new option for project
pipelines. With this set, if any one of the voting jobs fails, zuul
will report a build failure immediately and abort the remaining
jobs.

This patch enables fail-fast on the gate pipeline as any one
failing job means the patch will not merge.

[1] https://review.opendev.org/652764

Change-Id: I5e8cbe1f4eafd18e67b5b6914f5d459fb157a3a1