Fixed down server issue after reloading keepalived
When removing a UDP health monitor, keepalived is reloaded with a
configuration without any checkers.
But if keepalived has previously detected a down server, the state of
the server is unchanged and it will never be added to the list of IPVS
servers.

Restarting keepalived on configuration change works around this issue.

This issue is fixed in keepalived (>=2.0.14):
https://github.com/acassen/keepalived/issues/1163

Story: 2005774
Task: 33491

Change-Id: Iaa34db6cb1dfed98e96a585c5d105e263c7efa65
(cherry picked from commit 0a9f587015176723768a8a861265fda797c5d52e)