Add the Amphora image building guide to the docs
This patch adds the Amphora image building guide from the
diskimage-create README.rst to the Administration documentation.
It also re-organizes the Adminstration guides to be broken down
by category as the old page was becoming a long list of guides.
(I like that kind of problem)
The diskimage-create README has a few formating corrections to make
it render better for the documentation.

Change-Id: Ice4071e1f872c8c0d0595427cff6f02ffbcf7968