Prevent UDP LBs to use different IP protocol versions in amphora driver
The amphora doesn't support mixing IPv4 and IPv6 addresses for its
members and its VIP when using UDP load balancers.
This commit adds a validation step in the member_create and th e
member_batch_update functions of the amphora driver to ensure that IP
protocol versions are the same in a UDP load balancer.

Story: 2005876
Task: 33689

Change-Id: If6fb3fde9b43ac82af46eaddc48ec7a3a5b95602