Switch the default Ubuntu kernel to -virtual
A previous patch changed the default Ubuntu kernel for the amphora
images to the -kvm kernel[1]. This kernel is missing a module required
for cryptsetup.
This patch switches the default to the -virtual kernel, which is
slightly larger than the -kvm kernel, but includes the required
kernel module for cryptsetup.

[1] https://review.opendev.org/665861

Change-Id: Ia05d7e9f8ac2692e5769648255ef0e4e10a18c03