Lookup interfaces by MAC directly
Currently the amphora agent will lookup interfaces using the
interface name determined earlier in the plug method. This can
lead to a race condition with the udev interface renaming rule.
This patch changes the interface lookup to use the MAC address
directly and not rely on the interface name.

Story: 2006300
Task: 36013

Change-Id: I5bc21d5abdeb67a3a8ae88456735643463f15694