Fix listener deletion in ACTIVE/STANDBY topology
When removing listeners, listeners are removed from the load balancer's
listener list just before reconfiguring each amphora.
In case of ACTIVE/STANDBY topology, the code is performed on both
amphorae, so the listener is removed twice from the list.
This commit ensures that we don't remove an already removed listener.

Story: 2006329
Task: 36065

Change-Id: I426255f587f36b415eb999a9eb28cf0f91de94b0