Clarify that an example is not an actual list
In the flavor documentation there is an example output of the provider
capabilities list output. It is only an example and not an actual list
of available capabilities (this is why the command and API exit).
This patch adds "..." to the example to indicate it's not a complete
list as the sentence above the example states.

Change-Id: Ic2367ffbe5f116b7e386552dfc1fd17aa8235883
Story: 2006340
Task: 36090