Delete the periodic Fedora 28 amphora image job
The job broke around mid June and has not been fixed since that time.
Plus, Fedora 28 is no longer supported by the Fedora project and is not
tested by any of our dsvm scenario jobs. It is time to delete the job
from the periodic queue.

The community is more than welcome to help restore Fedora-based amphora
image builds. One might be interested in picking help change
https://review.opendev.org/#/c/673173/ where the build was successful
but was did not pass the dsvm scenario job.

Change-Id: Ic07629adb9419c74d5ed957a0640fb1dc8253521