Fix log offload file permissions in CentOS devstack
CentOS sets a restrictive umask by default that will override the
rsyslog file permissions settings. This can cause our tempest tests
to no be able to read the log offload files in our CentOS gate jobs.
This patch clears this umask override in the devstack rsyslog
configuration.

Note: This is a global setting for rsyslog and you probably do not
      want to set this outside of a testing environment.

Change-Id: I5962e335acff79152fa49c27c883aef2879593db