Allow the Octavia wsgi to accept argv parameters
Currently we only allow argv parameters to be passed into our wsgi
app. This means that some wsgi wrappers (such as pbr's script)
may not allow the octavia API to accept parameters such as
--config-file.
This patch attempts to pull in the sys.argv if none was passed into
the wsgi setup_app method.

Change-Id: I67cf32aac8aa8d8123ec8ca1ab433ff23fb1ffa2