Cap hacking version to <2
Hacking received a major update to version 2.0.0 that added new checks
that fail on stable branches. Instead of addressing those new check
errors on stable branches, it is better to cap hacking to <2. This way
we'll not see lint errors again should new checks be added in the
future.

Support for hacking 2.0.0 in master branch is being worked in
https://review.opendev.org/#/c/699302/

Since this patch touches requirements file, the requirements-check CI
job also runs and does not find matching gunicorn and sphinx
requirements. This patch fixes that too.

Change-Id: I4706972095d42815a3f3cec0bc5aa82a34922bf3