Change default tox envs from python37 to python3
Python versions supported by OpenStack change over time, and for minor
versions of Python 3 it is tedious to keep this file updated.

Since this does not impact zuul jobs in any way, nor prevent local
tests against py37, it should be safe to simply make this more easily
compatible for users that don't care about the specific Python versions
and just need basic tests to run.

The *only* thing this does is changes the default versions tested if none
are explicitly provided with `-e`.

Change-Id: I2372178351e961eeed5a819f39e75d54ba148798
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>