Add a periodic fast forward upgrade job
Fast forward upgrade (FFU) is an upgrade strategy that does in-place
upgrade and skips intermediary releases (e.g. N -> N+3).

This new periodic job fast forwards a cloud from Stein to master.

Please note that ideally a distro upgrade should also be performed but
Grenade does not support that today. For example, Bionic is a tested
Linux distribution on Stein, whereas on Victoria it is Focal.

Change-Id: I792f42468c37b6849c8fc085172841434d6a1376