Fix invalid DOWN status when updating a UDP pool
When adding a new UDP member or a UDP-CONNECT health-monitor to a UDP
pool, there can be a race condition in the first heartbeat message
sent to the health-manager service.
This message might contain a DOWN status for a working member that
hasn't been checked yet.

This commit introduces a new member status between the amphora-agent and
the health-manager: it indicates that the UDP pool has been updated and
that the status of a member is a transitional state, preventing an
incorrect ERROR status.

Story: 2007792
Task: 40042

Change-Id: Id9e19375ebca6a720e6a85006f5e8948d3aed760