Set octavia-grenade job non-voting
A recent uWSGI release broke devstack master and stable branches. It's
becoming increasingly difficult to maintain branches in Extended
Maintenance, especially the grenade job which depends on an older
release. The OpenStack QA team has decided [1] that the best thing to do
is to only perform upgrade testing between the current source contents
of adjacent Maintained or Development status branches.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-June/015499.html

Change-Id: Ifcb6ad47e8a69db4c059d72b240d1750a0c5ff38