Add non-voting octavia-v2-dsvm-scenario-centos-8
Recent change [1] improved the reliability and lowered the job duration
significatly (as low as 41m39s per a recent run) [2]. This enables us to
more comfortably (less wait time) and confidently (less timeouts) run
scenario tests on CentOS 8.

Should this job prove to have a record of being stable for a reasonable
amount of time/builds, we can follow up with a patch to promote it to
voting.

[1] https://review.opendev.org/#/c/738246/
[2] https://zuul.openstack.org/builds?job_name=octavia-v2-dsvm-scenario-centos-8

Change-Id: Ia20caacd31ae5a411981a33d9a7f2a6330902760