Set amphora-agent build-only packages
A recent change in DIB [1] broke our CentOS 8 amphora image builds only
because we've not been handling package dependencies propertly.

The DIB change now installs package "python3" which provides
/usr/bin/python3 and is a requirement of scripts in later build phases.
It so happens that we are removing package "python3-pip", a dependency
of "python36" too early so builds fail when bin python3 is required.

DIB package-installs element provides property "build-only". When set,
it will cause packages to be added both to the list of packages
installed and list of packages to be uninstalled. This allows expressing
build-time dependencies that should not end up in the final image as is
the case of packages like "build-essential" and "libffi-dev" in our
amphora-agent element.

As for the removal of Git, we do not explicitely install it so we should
not uninstall it either. Git is a fake dependency of source-repositories
element and will be abstent in the final image with [2].

[1] https://review.opendev.org/#/c/741868/
[2] https://review.opendev.org/#/c/745678/

Depends-On: https://review.opendev.org/#/c/745678/

Change-Id: I3b82180c8fcf2e2f053c6bcce405f1ea730f62ba
(cherry picked from commit fa1c5ef6dd15ce81c4a169307b7ab9970e9052d2)