Ensure xtrace is enabled in devstack plugin
Currently some phases of the devstack plugin do not have xtrace
enabled, which can make it hard to debug issues in the devstack
plugin.
This patch makes sure that xtrace is enabled while the Octavia
devstack plugin is running so that we can see the commands being
executed. It will restore the previous setting after our plugin
is done running.

Change-Id: Id6828b86779f1daca6a00a03f43c78fe26828f4f