Fix the lower-constraints job for focal
The move to focal for the tox jobs has raised issues with package
versions supporting python 3.8.
This patch bumps the version for the following packages:
PyMySQL==0.8.0
pyroute2==0.5.13
SQLAlchemy==1.2.19
cffi==1.14.0
cryptography==3.0
pyOpenSSL==19.1.0
oslo.db==8.30
oslo.messaging==12.4.0

This patch makes the grenade job non-voting as there is a chain
of gate fix patches required to fully fix all of the issues.

Change-Id: I95cc182a2dbcb2ea872e822ce9372bf5d7df63bd