Fix amphora failover when VRRP port is missing
In an amphora failover situation, if the VRRP port of the other
amphora in an active/standby pair is missing from neutron, the amphora
failover may fail with a "port not found" error.
This patch corrects this issue by allowing the amphora failover in
progress to complete even if the other amphora is also in a failed
state.

Train-Only: this commit also fixes the order of parameters in
v1/.../amphora_driver_tasks.py. It was fixed on other stable branches in
"Update amphora v2 for the failover refactor" (From
https://review.opendev.org/c/openstack/octavia/+/738292).

Story: 2008127
Task: 40851

Change-Id: I806d220236ad741b638ad9315537334f2d923031
(cherry picked from commit 66f3a63f442bb790dea5c529aa56a462d9515363)