Revert pyroute2 workaround
The pyroute2 module had a regression in 0.5.13, but neutron needed a fix
that was included in 0.5.13. It was easier for us to work around it than
for neutron to revert the update to 0.5.13, so we hacked a workaround[1].

A new pyroute2 release, 0.5.14, has been published, so this patch reverts
the hack and bumps our minimum version of pyroute2 to 0.5.14.

[1] https://review.opendev.org/#/c/744045

Change-Id: Ia4357a4f5db07e1a6b0beb11b58ce125215eeef1