Fix image build when installing agent from package
There is a bug in the octavia-lib DIB element where it runs install
actions that should only be ran when the amphora-agent is installed from
source. This issue causes the amphora image build to fail when
installing the amphora-agent from distribution packages (option -p in
the diskimage-create.sh tool).

Task: 40870
Story: 2008138

Change-Id: I46ba7798322f24255ab7d49db4b4df886e92cbf3