Followup Remove haproxy_check_script for UDP-only
The original patch: Ib2a80afb9223da2c0af6f6add250ffaaa20a316f

First attempt at fixing this issue assumed that the update function
would actually be called again after the initial LB creation, which it
is not. We should ensure that this is called when we do any UDP listener
creation, so the file can be removed if necessary.

Change-Id: Iec1ec7df420e3746b47004c8c0374d2d8274f45b