Fix python2 AttributeError with strptime
Fix a potential python2 error when calling strptime, the first call to
strptime is not thread safe and might trigger an exception.
A workaround was added by calling strptime before creating a thread but
the caller is itself within a thread, so it may fail.

The workaround is now to explicitly import _strptime as described in the
upstream python issue https://bugs.python.org/issue7980

Note: This commit only applies to branches that support python2 (initial
commit in stable/train)

Story 2008211
Task 40997

Change-Id: I0e7119d5a379f8662c4913d7232ef28535a29b32