Fix load balancers with failed amphora failover
This patch is a complement of [1] which we ensure
a fresh lb object, but only for AmpListenersUpdate
class.

It was observed, at least on train and ussuri
deployments that multiple failed amphoras were still
happening when using session persistence.

That is fixed and failover works flawless with session
persistence when we also ensure a fresh lb object on
AmphoraIndexListenerUpdate class.

[1] https://review.opendev.org/#/c/756597/

Change-Id: I923accd73e0c9cadc91c115157c576432f428622
Story: 2008099
Task: 40802