Validate listener protocol in amphora driver
Validate that the amphora driver supports the listener protocol when
receving a listener_create request.

It returns an UnsupportedOptionError exception to the user if the
amphora driver doesn't support a protocol that is defined in the API
(ex: SCTP is supported in the API in the Victoria release, but not in
the amphora driver).

Story: 2008545
Task: 41647

Change-Id: I1c5cb987945a7a465bfecfda399dfe93fc1b76bb