Fix amphora image build
DIB dropped Python 2 support in its version 3.0.0 major release and as
so we had to pin it to 2.30.0 in stable/stein and stable/train branches.
However, upstream get-pip.sh recently introduced a regression where it
does not work well with Python <-3.5. Sadly, DIB 2.30.0 points to master
get-pip.sh [1] and DIB is branchless so the fix cannot be backported and
we are in a deadlock scenario where Python <=3.5 amphora images can't be
built.

To work-around this issue on stable/train, we have to make a compromise:
either build amphora images using master DIB and Python 3 only or be
stuck (unless someone else can come up with another approach, all ears).
Ussuri and newer releases are not impacted by this issue as they are
Python >=3.6 only releases.

The same issue also applies to Stein and that has an impact on the
octavia-grenade CI job in stable/train. Since Stein Octavia is EOL [2],
we no longer have to test upgrades.

[1] https://review.opendev.org/c/openstack/diskimage-builder/+/772254
[2] https://review.opendev.org/c/openstack/releases/+/772847

Change-Id: I13da9a49ace803e957dc52e4882b78141bc7bef3