Fix amphora image build
Upstream get-pip.sh recently introduced a regression where it does not
work well with Python <=3.5. DIB, upper-constrained at version 2.35.0 in
requirements/upper-constraints.txt, wants to install pip with Python 2,
and that fails with with error message:

  "ERROR: This script does not work on Python 2.7 The minimum supported
   Python version is 3.6."

See attached story for more logging.

To work-around this issue on stable/ussuri, we have to make a
compromise: build amphora images using a >= 3.x DIB version. We could
either pick a 3.x DIB released version or Git master branch. For
consistency with what was done in stable/stein and stable/train
not so long ago, we propose Git master branch.

Task: 41991
Story: 2008688

Change-Id: I4531ff80ac82a61e527c024101e0785d49aa135b