Fix incorrect ERROR status with IPv6 UDP members
Some IPv6 UDP members were incorrectly marked in ERROR status because of
a formatting issue between the keepalived configuration file and the
ipvsadm output. Both are used to compute the state of the members and
when a member's address contained '*:0:*', parsing was incorrect. Now
the health message generation function uses only the compressed IPv6
notation instead of mixing notations.

Story: 2008604
Task: 41783

Change-Id: I2fe94cd4c000f143c59c69e82d03c690acf5e0c3