Fix weighted round-robin about UDP and SCTP listener
'lb_algo rr' in keepalived won't work correctly with weight
for UDP and SCTP listeners, it should be 'lb_algo wrr'.

'wrr' is superset of 'rr' algorithm, as round-robin is
a specical instance of the weighted round-robin scheduling,
in which all the weights are equal. [1]

Algorithm in HAProxy is set to 'roundrobin', which also support
weights, but in keepalived must be set to 'wrr' to work with
the weighted round-robin scheduling, as it is different to 'rr'.

[1] https://www.keepalived.org/doc/scheduling_algorithms.html

Story 2008462
Task 41491

Change-Id: Ic63929d8864e5285baf70dd85e6362988bf2863f