Fix grenade resource waiters
The grenade job resource.sh script is waiting for the created object
status but not the overall load balancer status to go ACTIVE.
This can lead to the script failing with a 409 error.

This patch adds a check for the load balancer to be unlocked before
advancing to the next create call.

Change-Id: I6505243ddbf1eab7d110e9bfa03bffda840f07ae