Don't fail if a provider driver cannot be loaded in Octavia API
Fix an issue that prevents the Octavia API service to be correctly
initialized when it fails to load a provider driver.

When enabled_provider_drivers setting is not correctly configured (ex:
if it contains a non existing driver) or when a provider driver fails
to load, the exception was not caught by the Octavia API service, so the
service was not properly configured and the whole Octavia API was
unreachable.

Now the Octavia API service skips the driver initialization in case of
errors, removes the driver(s) from the enabled list, and the other
provider drivers are functional.

Story 2008710
Task 42044


Change-Id: I34341e1aaad1524e3e0834309c5f6897f176af53
Signed-off-by: Zachary Raines <zachary.raines@canonical.com>
Signed-off-by: Michael Johnson <johnsomor@gmail.com>
(cherry picked from commit 6fc1abb780a2fc1413d5b206cec4b033a638c3c4)