Fix pool ALPN compatibility with older amphora
The ALPN for pools patch introduced a bug that causes amphora to
fail if the HAProxy version running inside the amphroa does not
support ALPN on backend members (bionic images for example).
This patch adds compatibility support for those older HAProxy versions
by removing the ALPN configuration settings if the HAProxy version
is too old to support the functionality.

Story: 2008780
Task: 42171
Change-Id: Iac949431f0901aaa0a60e5c39c2aab44cb9c1970